Core/Client - Fix creating alias/NFT addresses from OutputID and use new JSON formatting rules (#379)
* get_txn_included_txn_id_len_fix


* 1. TRANSACTION_ESSENCE_TYPE = 1
2. CORE_MESSAGE_PAYLOAD_TRANSACTION = 6


* Hex encoded strings has 0x prefix in JSON format, uint64_t numbers are represented as strings in JSON format

* Added CORE_MESSAGE_PAYLOAD_DEPRECATED,


* Hex encoded strings has 0x prefix in JSON format in unlock_blocks

* Fix some unit tests to be compatible with new JSON format

* Fix some unit tests to be compatible with new JSON format

* Fix some unit tests to be compatible with new JSON format

* Fix get included message API call

* Fix deserializing metadata and tag feature blocks. Fix some unit tests.

* Fix serializing unlock addresses for JSON message

* Fix transaction essence number in serialization for JSON message

* Fix faucet enqueue unit test

* Fix unit test for node info API call

* Remove deprecated unit test file.

* Removed deprecated API endpoint

* Fix bug in alias_address_from_output and nft_address_from_output

* Added alias address as query parameter for API calls

* Fixed some JSON serialization/deserialization

* Refactor checking for 0x prefix and adding 0x prefix

* Refactored tag and metadata JSON deserializations

* Check of 0x in hex string when gathering message parents

* Fix GCC warning

* Fix GCC warning

* Fix another GCC warning. pruning_milestone_index is always bigger or equal zero

* Added missing function return value checks

* Refactored hex_2_bin and bin_2_hex to use additional argument for prefix.

* Use macro for "0x" prefix

* Fixed foundry output JSON format

* Fix deserialization for tips and message children.

* Fix sending transaction and tagged message

+++ src/client/api/restful/faucet_enqueue.h
@@ -18,7 +18,7 @@
 typedef struct {
   char bech32_address[BIN_TO_HEX_STR_BYTES(ED25519_PUBKEY_BYTES)];  ///< The bech32 encoded address that is
                                                                     ///< returned in response
-  uint64_t waiting_reqs_count;                                      ///< The number of requests in faucet queue
+  uint32_t waiting_reqs_count;                                      ///< The number of requests in faucet queue
 } faucet_enqueue_t;
 
 /**
