+++ src/core/models/payloads/milestone.h
@@ -19,7 +19,7 @@
 typedef struct {
   uint32_t type;       // pyalod type, set to 1 denotes a milestone pyalod
   uint32_t index;      // The index number of the milestone.
-  uint64_t timestamp;  // The Unix time (seconds since Unix epoch) at which the milestone was issued.
+  uint32_t timestamp;  // The Unix time (seconds since Unix epoch) at which the milestone was issued.
   UT_array *parents;   // Parents of milestone message.
   byte_t inclusion_merkle_proof[CRYPTO_BLAKE2B_256_HASH_BYTES];  // The Merkle tree hash (BLAKE2b-256) of the message
                                                                  // IDs of all the not-ignored state-mutating
